[api/workspace] Enhancement lint + remove update + separation of hooks
