+++ api/src/services/workspace/workspace.service.ts
@@ -3,7 +3,8 @@ import { createSwaggerServiceOptions } from 'feathers-swagger'
 import type { Application } from '../../declarations'
 import { ObjectionAdapterOptions } from '../../feathers-objection'
 
-import { WorkspaceService, workspaceHooks } from './workspace.class'
+import { WorkspaceService } from './workspace.class'
+import { workspaceHooks } from './workspace.hooks'
 import { WorkspaceModel } from './workspace.model'
 import { workspaceDataSchema, workspaceQuerySchema, workspaceSchema } from './workspace.schema'
 
@@ -20,7 +21,7 @@ export function workspaceService(app: Application): void {
   // Register our service on the Feathers application
   app.use(API_PATH.WORKSPACE.ROOT, new WorkspaceService(options), {
     // A list of all methods this service exposes externally
-    methods: ['find', 'get', 'create', 'update', 'patch', 'remove'],
+    methods: ['find', 'get', 'create', 'patch', 'remove'],
     // You can add additional custom events to be sent to clients here
     events: [],
     docs: createSwaggerServiceOptions({
